+++ .circleci/config.yml
@@ -31,7 +31,8 @@ unit-config: &unit-config
         name: Install PHP extensions
         command: |
             printf '\n' | sudo pecl install redis
-            sudo docker-php-ext-enable redis
+            printf '\n' | sudo pecl install apcu
+            sudo docker-php-ext-enable redis apcu
 
     - run:
         name: Install composer packages
@@ -49,7 +50,13 @@ unit-config: &unit-config
 
     - run:
         name: Wait for backing services to start
-        command: dockerize -wait tcp://127.0.0.1:6379 -timeout 30s
+        command: dockerize -wait tcp://127.0.0.1:6379 -wait tcp://127.0.0.1:9091 -timeout 30s
+
+    - run:
+        name: Export environment variables
+        command: |
+          echo 'export PUSH_GATEWAY_ADDR="127.0.0.1:9091"' >> $BASH_ENV
+          source $BASH_ENV
 
     - run:
         name: Run PHP Code Sniffer
@@ -57,7 +64,7 @@ unit-config: &unit-config
 
     - run:
         name: Run tests
-        command: ./vendor/bin/phpunit
+        command: php -d apc.enable_cli='On' ./vendor/bin/phpunit
 
 version: 2.1
 jobs:
@@ -66,6 +73,7 @@ jobs:
     docker:
       - image: circleci/php:7.4-cli-node
       - image: circleci/redis:5-alpine
+      - image: prom/pushgateway
     environment:
       COMPOSER_PREFER_LOWEST: 0
 
@@ -74,6 +82,7 @@ jobs:
     docker:
       - image: circleci/php:7.4-cli-node
       - image: circleci/redis:5-alpine
+      - image: prom/pushgateway
     environment:
       COMPOSER_PREFER_LOWEST: 1
 
@@ -82,6 +91,7 @@ jobs:
     docker:
       - image: circleci/php:7.3-cli-node
       - image: circleci/redis:5-alpine
+      - image: prom/pushgateway
     environment:
       COMPOSER_PREFER_LOWEST: 0
 
@@ -90,6 +100,7 @@ jobs:
     docker:
       - image: circleci/php:7.3-cli-node
       - image: circleci/redis:5-alpine
+      - image: prom/pushgateway
     environment:
       COMPOSER_PREFER_LOWEST: 1
