+++ composer.json
@@ -19,7 +19,7 @@
         "symfony/polyfill-apcu": "^1.6"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.4",
+        "phpunit/phpunit": "^8.4.3",
         "squizlabs/php_codesniffer": "^3.5"
     },
     "suggest": {
@@ -32,8 +32,8 @@
         }
     },
     "autoload-dev": {
-        "psr-0": {
-            "Test\\Prometheus\\": "tests/"
+        "psr-4": {
+            "Prometheus\\Tests\\": "tests/"
         }
     },
     "extra": {
