+++ examples/metrics.php
@@ -4,17 +4,19 @@ require __DIR__ . '/../vendor/autoload.php';
 
 use Prometheus\CollectorRegistry;
 use Prometheus\RenderTextFormat;
+use Prometheus\Storage\APC;
+use Prometheus\Storage\InMemory;
 use Prometheus\Storage\Redis;
 
 $adapter = $_GET['adapter'];
 
 if ($adapter === 'redis') {
     Redis::setDefaultOptions(['host' => $_SERVER['REDIS_HOST'] ?? '127.0.0.1']);
-    $adapter = new Prometheus\Storage\Redis();
+    $adapter = new Redis();
 } elseif ($adapter === 'apc') {
-    $adapter = new Prometheus\Storage\APC();
+    $adapter = new APC();
 } elseif ($adapter === 'in-memory') {
-    $adapter = new Prometheus\Storage\InMemory();
+    $adapter = new InMemory();
 }
 $registry = new CollectorRegistry($adapter);
 $renderer = new RenderTextFormat();
