+++ tests/BlackBox/BlackBoxTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Test;
+namespace Prometheus\Tests\BlackBox;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\Promise;
@@ -18,9 +18,13 @@ class BlackBoxTest extends TestCase
      */
     private $adapter;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->adapter = getenv('ADAPTER');
+        if (empty($this->adapter)) {
+            $this->markTestSkipped("No value for 'ADAPTER' environment variable.");
+        }
+
         $this->client = new Client(['base_uri' => 'http://nginx:80/']);
         $this->client->get('/examples/flush_adapter.php?adapter=' . $this->adapter);
     }
@@ -43,7 +47,7 @@ class BlackBoxTest extends TestCase
         echo "\ntime: " . ($end - $start) . "\n";
 
         $metricsResult = $this->client->get('/examples/metrics.php?adapter=' . $this->adapter);
-        $body = (string)$metricsResult->getBody();
+        $body = (string) $metricsResult->getBody();
         echo "\nbody: " . $body . "\n";
         $this->assertThat(
             $body,
@@ -73,7 +77,7 @@ class BlackBoxTest extends TestCase
         echo "\ntime: " . ($end - $start) . "\n";
 
         $metricsResult = $this->client->get('/examples/metrics.php?adapter=' . $this->adapter);
-        $body = (string)$metricsResult->getBody();
+        $body = (string) $metricsResult->getBody();
 
         $this->assertThat($body, $this->stringContains('test_some_counter{type="blue"} ' . $sum));
     }
@@ -102,7 +106,7 @@ class BlackBoxTest extends TestCase
         echo "\ntime: " . ($end - $start) . "\n";
 
         $metricsResult = $this->client->get('/examples/metrics.php?adapter=' . $this->adapter);
-        $body = (string)$metricsResult->getBody();
+        $body = (string) $metricsResult->getBody();
 
         $this->assertThat($body, $this->stringContains(<<<EOF
 test_some_histogram_bucket{type="blue",le="0.1"} 1
