+++ tests/Unit/AbstractHistogramTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Test\Prometheus;
+namespace Prometheus\Tests\Unit;
 
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
