+++ tests/Unit/InMemory/CounterTest.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Test\Prometheus\InMemory;
+namespace Prometheus\Tests\Unit\InMemory;
 
 use Prometheus\Storage\InMemory;
-use Test\Prometheus\AbstractHistogramTest;
+use Prometheus\Tests\Unit\AbstractCounterTest;
 
 /**
  * See https://prometheus.io/docs/instrumenting/exposition_formats/
  */
-class HistogramTest extends AbstractHistogramTest
+class CounterTest extends AbstractCounterTest
 {
 
     public function configureAdapter()
