+++ tests/Unit/Redis/CounterWithPrefixTest.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Test\Prometheus\Redis;
+namespace Prometheus\Tests\Unit\Redis;
 
 use Prometheus\Storage\Redis;
-use Test\Prometheus\AbstractGaugeTest;
+use Prometheus\Tests\Unit\AbstractCounterTest;
 
 /**
  * See https://prometheus.io/docs/instrumenting/exposition_formats/
  * @requires extension redis
  */
-class GaugeWithPrefixTest extends AbstractGaugeTest
+class CounterWithPrefixTest extends AbstractCounterTest
 {
     public function configureAdapter()
     {
