+++ tests/Unit/Redis/GaugeTest.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Test\Prometheus\Redis;
+namespace Prometheus\Tests\Unit\Redis;
 
 use Prometheus\Storage\Redis;
-use Test\Prometheus\AbstractHistogramTest;
+use Prometheus\Tests\Unit\AbstractGaugeTest;
 
 /**
  * See https://prometheus.io/docs/instrumenting/exposition_formats/
  * @requires extension redis
  */
-class HistogramTest extends AbstractHistogramTest
+class GaugeTest extends AbstractGaugeTest
 {
     public function configureAdapter()
     {
