+++ tests/bootstrap.php
@@ -1,15 +1,9 @@
 <?php
+
 error_reporting(-1);
 date_default_timezone_set('UTC');
-$autoload = __DIR__ . '/../vendor/autoload.php';
-if (!file_exists($autoload)) {
-    echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~" . PHP_EOL;
-    echo " You need to execute `composer install` before running the tests. " . PHP_EOL;
-    echo " Vendors are required for complete test execution. " . PHP_EOL;
-    echo "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~" . PHP_EOL . PHP_EOL;
-    exit(1);
-}
-$loader = require $autoload;
-$loader->add('Test\\Prometheus', __DIR__);
+
+require __DIR__ . '/../vendor/autoload.php';
 
 define('REDIS_HOST', isset($_SERVER['REDIS_HOST']) ? $_SERVER['REDIS_HOST'] : '127.0.0.1');
+define('PUSH_GATEWAY_ADDR', isset($_SERVER['PUSH_GATEWAY_ADDR']) ? $_SERVER['PUSH_GATEWAY_ADDR'] : 'pushgateway:9091');
